Fix loading of user data after groups are added
The API being used to refresh the logged in user's data after a group
was added returns data for all users instead of the current user as
the UI code was expecting. This caused bad data to be stored in
localstorage and broke the user's session (requiring the user log in
again to fix).

A simpler API is now used which just returns the details for a single
user.